Complete migration
