+++ trakt/trakt.go
@@ -8,7 +8,6 @@ import (
 	"net/http"
 	"time"
 
-	"github.com/marcus-crane/gunslinger/models"
 	"github.com/marcus-crane/gunslinger/playback"
 	"github.com/marcus-crane/gunslinger/utils"
 )
@@ -19,7 +18,59 @@ var (
 	tmdbEpisodeEndpoint  = "https://api.themoviedb.org/3/tv/%d/season/%d/episode/%d/images"
 )
 
-func getArtFromTMDB(apiKey string, traktResponse models.NowPlayingResponse) (string, error) {
+type NowPlayingResponse struct {
+	ExpiresAt      string       `json:"expires_at"`
+	StartedAt      string       `json:"started_at"`
+	Action         string       `json:"action"`
+	Type           string       `json:"type"`
+	Movie          TraktSummary `json:"movie"`
+	Episode        TraktEpisode `json:"episode"`
+	Show           TraktSummary `json:"show"`
+	PodcastEpisode TraktEpisode `json:"podcast_episode"`
+	Podcast        TraktSummary `json:"podcast"`
+}
+
+type TraktEpisode struct {
+	Season        int      `json:"season"`
+	Number        int      `json:"number"`
+	Title         string   `json:"title"`
+	IDs           TraktIDs `json:"ids"`
+	Overview      string   `json:"overview"`
+	OverviewPlain string   `json:"overview_plain"`
+	Explicit      bool     `json:"explicit"`
+	Runtime       int      `json:"runtime"`
+}
+
+type TraktSummary struct {
+	Title         string   `json:"title"`
+	Year          int      `json:"year"`
+	IDs           TraktIDs `json:"ids"`
+	Overview      string   `json:"overview"`
+	OverviewPlain string   `json:"overview_plain"`
+	Author        string   `json:"author"`
+	Homepage      string   `json:"homepage"`
+}
+
+type TraktIDs struct {
+	Trakt int    `json:"trakt"`
+	Slug  string `json:"slug"`
+	TVDB  int    `json:"tvdb"`
+	IMDB  string `json:"imdb"`
+	TMDB  int    `json:"tmdb"`
+	Apple int    `json:"apple"`
+}
+
+type TMDBImageResponse struct {
+	ID      int         `json:"id"`
+	Stills  []TMDBImage `json:"stills"`
+	Posters []TMDBImage `json:"posters"`
+}
+
+type TMDBImage struct {
+	FilePath string `json:"file_path"`
+}
+
+func getArtFromTMDB(apiKey string, traktResponse NowPlayingResponse) (string, error) {
 	url := ""
 	if traktResponse.Type == "movie" {
 		url = fmt.Sprintf(tmdbMovieEndpoint, traktResponse.Movie.IDs.TMDB)
@@ -38,7 +89,7 @@ func getArtFromTMDB(apiKey string, traktResponse models.NowPlayingResponse) (str
 	if err != nil {
 		return "", err
 	}
-	var tmdbImageResponse models.TMDBImageResponse
+	var tmdbImageResponse TMDBImageResponse
 
 	if err = json.Unmarshal(body, &tmdbImageResponse); err != nil {
 		slog.Error("Failed to fetch image data from TMDB", slog.String("body", string(body)))
@@ -115,7 +166,7 @@ func GetCurrentlyPlayingTrakt(ps *playback.PlaybackSystem, client http.Client) {
 		)
 		return
 	}
-	var traktResponse models.NowPlayingResponse
+	var traktResponse NowPlayingResponse
 
 	if err = json.Unmarshal(body, &traktResponse); err != nil {
 		// TODO: Check status code
